Move generate_bin_centers to generators
